+++ interface/menus/menu_define.py
@@ -1,0 +1,8 @@
+### DEFINITION OF MENUS USING A NESTED DICT
+### TOP LEVEL DICT IS LIST OF SUBMENUS
+### NEXT LEVEL DICT KEYS ARE SUBMENU OPTIONS
+### NEXT LEVEL DICT VALUES ARE NAMES OF FUNCTIONS TO CALL IN log/api/LogicApi
+
+menu_definition = {
+    "Cadets": {"View Master List of Cadets": "view_master_list_of_cadets"}
+}