+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Release notes
 
-## Version 0.01
+## Version 0.01 (2023-09-13)
+
+- Created skeleton CLI and CSV app which can display a list of cadets from an example .csv
+- Created web interface object, but with no functionality